fix: divisa logica dall'implementazione
